requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
appdirs==1.4.4
async-timeout==4.0.2
attrs==22.2.0
Bottleneck==1.3.5
brotlipy==0.7.0
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.4
click==8.1.3
colorama==0.4.6
contourpy==1.0.5
cryptography==39.0.1
cycler==0.11.0
dataclasses-json==0.5.7
discord.py==2.2.2
dnspython==2.3.0
docker-pycreds==0.4.0
et-xmlfile==1.1.0
fonttools==4.25.0
frozenlist==1.3.3
gitdb==4.0.10
GitPython==3.1.31
google-search-results==2.4.2
gptcache==0.1.21
greenlet==2.0.1
idna==3.4
joblib==1.2.0
kiwisolver==1.4.4
langchain==0.0.150
loguru==0.7.0
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.7.1
multidict==6.0.2
munkres==1.1.4
mypy==0.981
mypy-extensions==0.4.3
numexpr==2.8.4
numpy==1.24.3
openai==0.27.4
openapi-schema-pydantic==1.2.4
openpyxl==3.0.9
packaging==23.1
pandas==1.5.3
pandas-stubs==2.0.0.230412
pathtools==0.1.2
Pillow==9.4.0
pinecone-client==2.2.1
pip==23.0.1
platformdirs==3.5.0
plotly==5.14.1
pooch==1.7.0
protobuf==3.20.3
psutil==5.9.0
pycparser==2.21
pydantic==1.10.2
pyOpenSSL==23.1.1
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
regex==2023.3.23
requests==2.29.0
scikit-learn==1.2.2
scipy==1.10.1
sentry-sdk==1.21.1
setproctitle==1.2.2
setuptools==66.0.0
six==1.16.0
smmap==3.0.5
SQLAlchemy==1.4.39
stringcase==1.2.0
tenacity==8.2.2
threadpoolctl==3.1.0
tiktoken==0.3.3
toml==0.10.2
tomli==2.0.1
tqdm==4.65.0
types-pytz==2023.3.0.0
typing_extensions==4.5.0
typing-inspect==0.8.0
urllib3==1.26.15
wandb==0.15.0
wheel==0.38.4
yarl==1.8.1
